+++ CHANGELOG.md
@@ -25,7 +25,9 @@
 + [x] Lists: fix list style position for ordered list
 + [x] Dropdown: add service class `stay-open`. When an element has this class, an element can't be closed when the user clicks on the document.
 + [x] Collapse: change default animation duration to `100ms`
-+ [x] Colors: add accent colors `div`
++ [x] Colors: add accent colors to `div`
++ [x] Tiles: set cover default position to `center center`
++ [x] Tiles: add attribute for image slide `data-cover-position`
 
 ### 4.2.39
 + [x] NavView: minor improve styling